pyproject.toml
[project]
name = "Agente"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "datasets>=4.0.0",
    "opencv-python>=4.12.0.88",
    "tokenizer>=3.4.5",
    "torch>=2.7.1",
    "transformers[torch]>=4.54.1",
]